requirements.txt
flask==2.3.3
selenium==4.11.2
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
schedule==1.2.0
yagmail==0.15.293
playwright==1.49.1
pandas==2.2.3
APScheduler==3.10.4
